+++ plugin_base/src/plugin_base/plugin_base/gui/lnf.hpp
@@ -19,9 +19,9 @@ public juce::LookAndFeel_V4 {
 public:
   juce::Font font() const;
   gui_colors const& colors() const;
-  int combo_height() const { return _desc->plugin->gui.font_height + 4; }
-  int toggle_height() const { return _desc->plugin->gui.font_height + 4; }
-  lnf(plugin_desc const* desc, int custom_section, int module_section, int module);
+  int combo_height(bool tabular) const { return _desc->plugin->gui.font_height + (tabular ? 8 : 4); }
+  int toggle_height(bool tabular) const { return _desc->plugin->gui.font_height + (tabular ? 8 : 4); }
+  lnf(plugin_desc const* desc, int custom_section, int module_section, int module); 
 
   int getDefaultScrollbarWidth() override { return 8; }
   bool areScrollbarButtonsVisible() override { return true; }