+++ src/firefly_synth/firefly_synth/fx.cpp
@@ -106,8 +106,8 @@ dist_mode_items()
 {
   std::vector<list_item> result;
   result.emplace_back("{4216D379-72FD-4C2D-B594-20C175CF275E}", "No Filter");
-  result.emplace_back("{90657FC7-42E9-4D8E-88D1-5380F916A6C1}", "Filter To Shaper");
-  result.emplace_back("{01871DF5-5834-43F9-B95C-09D8102BF985}", "Shaper To Filter");
+  result.emplace_back("{90657FC7-42E9-4D8E-88D1-5380F916A6C1}", "Flt To Shaper");
+  result.emplace_back("{01871DF5-5834-43F9-B95C-09D8102BF985}", "Shaper To Flt");
   return result;
 }
 
@@ -646,7 +646,7 @@ fx_topo(int section, gui_colors const& colors, gui_position const& pos, bool glo
   svf_bottom.gui.bindings.visible.bind_params({ param_type }, [](auto const& vs) { return vs[0] == type_off || vs[0] == type_svf; });
   auto& svf_kbd = result.params.emplace_back(make_param(
     make_topo_info("{9EEA6FE0-983E-4EC7-A47F-0DFD79D68BCB}", true, "SV Filter Keytrack", "Keytrack", "SVF Kbd", param_svf_kbd, 1),
-    make_param_dsp_accurate(param_automate::modulate), make_domain_percentage(-2, 2, 1, 0, true),
+    make_param_dsp_accurate(param_automate::modulate), make_domain_percentage(-2, 2, global? 0: 1, 0, true),
     make_param_gui_single(section_svf_bottom, gui_edit_type::hslider, { 0, 0 },
       make_label(gui_label_contents::name, gui_label_align::left, gui_label_justify::center))));
   svf_kbd.gui.bindings.enabled.bind_params({ param_type }, [](auto const& vs) { return vs[0] == type_svf; });
@@ -666,7 +666,7 @@ fx_topo(int section, gui_colors const& colors, gui_position const& pos, bool glo
       make_label(gui_label_contents::name, gui_label_align::left, gui_label_justify::center))));
   comb_mode.gui.bindings.enabled.bind_params({ param_type }, [](auto const& vs) { return vs[0] == type_cmb; });
   comb_mode.gui.bindings.visible.bind_params({ param_type }, [](auto const& vs) { return vs[0] == type_cmb; });
-  comb_mode.info.description = "Selects the state-variable filter mode.";
+  comb_mode.info.description = "Selects the comb filter mode.";
   auto& comb_top = result.sections.emplace_back(make_param_section(section_comb_top,
     make_topo_tag_basic("{54CF060F-3EE7-4F42-921F-612F8EEA8EB0}", "Comb Filter Top"),
     make_param_section_gui({ 0, 1 }, { { 1 }, { 1, 1 } })));