+++ changelog.md
@@ -1,3 +1,11 @@
+### March 8, 2024 - V1.2.1.
+
+- Add big pad 2 preset.
+- Scale back default ui width to 1280.
+- Bugfix: osc graph scales to 100% * gain.
+- Bugfix: default global SVF keyboard tracking to 0.
+- UI scaling from 50% to 200% (was 100 to unlimited).
+
 ### March 6, 2024 - V1.2.0.
 
 - Lots of small ui fixes / cleanup.